backend/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-speech==2.0.1
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.17
MarkupSafe==1.1.1
mypy-extensions==0.4.3
nltk==3.4.5
packaging==20.9
proto-plus==1.14.2
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydub==0.24.1
pymongo==3.11.3
pyparsing==2.4.7
python-dotenv==0.15.0
pytz==2021.1
PyYAML==5.4.1
rake-nltk==1.0.4
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
Werkzeug==1.0.1
